+++ packages/decap-cms-core/src/reducers/__tests__/mediaLibrary.spec.js
@@ -7,7 +7,7 @@ import mediaLibrary, {
   selectMediaDisplayURL,
 } from '../mediaLibrary';
 
-jest.mock('uuid/v4');
+jest.mock('uuid');
 jest.mock('../entries');
 jest.mock('../');