+++ packages/decap-cms/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.7](https://github.com/decaporg/decap-cms/compare/decap-cms@3.0.6...decap-cms@3.0.7) (2023-09-06)
+
+**Note:** Version bump only for package decap-cms
+
+
+
+
+
+## [3.0.6](https://github.com/decaporg/decap-cms/compare/decap-cms@3.0.5...decap-cms@3.0.6) (2023-09-06)
+
+**Note:** Version bump only for package decap-cms
+
+
+
+
+
 ## [3.0.5](https://github.com/decaporg/decap-cms/compare/decap-cms@3.0.4...decap-cms@3.0.5) (2023-08-25)
 
 **Note:** Version bump only for package decap-cms